+++ src/app/index.d.ts
@@ -1,0 +1,8 @@
+import SagaTester from './sagaTester';
+import diffTwoObjects from './diffTwoObjects';
+import mockSelector from './mockSelector';
+import mockGenerator from './mockGenerator';
+import triggerSagaWithAction from './triggerSagaWithAction';
+import PLACEHOLDER_ARGS from './PLACEHOLDER_ARGS';
+
+export { SagaTester, diffTwoObjects, mockGenerator, mockSelector, triggerSagaWithAction, PLACEHOLDER_ARGS };