+++ src/app/index.js
@@ -3,4 +3,4 @@ export { default as mockGenerator } from './mockGenerator';
 export { default as mockSelector } from './mockSelector';
 export { default as diffTwoObjects } from './diffTwoObjects';
 export { default as triggerSagaWithAction } from './triggerSagaWithAction';
-export { default as PLACEHOLDER_ARGS } from './helpers/PLACEHOLDER_ARGS';
+export { default as PLACEHOLDER_ARGS } from './PLACEHOLDER_ARGS';